// client/src/types/Character.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BonusEntry } from "./BonusEntry";
import type { JsonValue } from "./serde_json/JsonValue";

export type Character = { id: string, campaign_id: string, owner_id: string, game_system_id: string, name: string, portrait_asset_id: string | null, visible_to_players: boolean, fields: { [key in string]: JsonValue }, bonuses: { [key in string]: Array<BonusEntry> }, created_at: string, updated_at: string, };